회원탈퇴 인증

+++ client/src/pages/Myinfo.js
@@ -122,16 +122,35 @@ class ModalDelete extends React.Component {
         this.state = {
             modal: false,
             deleteSuccess: false,
+            password: null,
         };
 
         this.toggle = this.toggle.bind(this);
     }
 
+    updatePassword = (e) => {
+        console.log(e.target.value);
+        this.setState({
+            password:e.target.value
+        })
+    }
     delete = async () => {
+        // const userid = e.target.elements.userid.value;
+        // const password = e.target.elements.password.value;
+    
+        const userid = this.props.userid;
+        const password = this.state.password;
+        let data = {userid, password};
+    
+        const confirmPassword = await axios.post('/mongo/login', data);
+        
+        if (!confirmPassword.data.result) {
+            alert("잘못된 비밀번호입니다\n다시 시도해 주세요!");
+            return;
+        }
+
         console.log(this.props.userid);
         const res = await axios.delete('/mongo/user/delete/' + this.props.userid);
-
-        
         if (res.data.result) {
             this.setState({
                 deleteSuccess: true
@@ -162,16 +181,36 @@ class ModalDelete extends React.Component {
 
                 <Button color="danger" onClick={this.toggle}>{this.props.buttonLabel} 회원 탈퇴하기</Button>
                 <Modal isOpen={this.state.modal} toggle={this.toggle} className={this.props.className}>
-                    <ModalHeader toggle={this.toggle}>탈퇴하기</ModalHeader>
+                    <ModalHeader toggle={this.toggle}>탈퇴안내</ModalHeader>
                     <ModalBody>
-                        <p>정말 탈퇴하시겠습니까??</p>
+                        <p>회웥탈퇴 진행에 앞서 아래의 사랑을 <strong>반드시 확인</strong>하시길 바랍니다</p>  
+                        <div className="notice">
+                            <p>1. 한번 탈퇴하면 ㅇㄹㅇㄹㅇㄹ</p>   
+                            <p>1. 한번 탈퇴하면 ㅇㄹㅇㄹㅇㄹ</p>  
+                            <p>1. 한번 탈퇴하면 ㅇㄹㅇㄹㅇㄹ</p>  
+                            <p>1. 한번 탈퇴하면 ㅇㄹㅇㄹㅇㄹ</p>  
+                            <p>1. 한번 탈퇴하면 ㅇㄹㅇㄹㅇㄹ</p>                     
+                            </div>
+
+                            <div className="reason">
+                                <p>탈퇴사유 입력</p>
+                                
+                                <p>향후 더 나은 서비스 제공을 위해 탈퇴사유를 입력하세요</p>
+                                <input></input>
+                            </div>
+
+                            <div className="password">
+                                <p>패스워드 인증</p>
+                            
+                                <p>타인에 의한 탈퇴를 방자하기 위한 추가 인증입니다</p>
+                                <input onChange={this.updatePassword} placeholder="패스워드를 입력하세요." name="password" type="password"></input>
+                            </div>
                     </ModalBody>
                     <ModalFooter>
                         <Button color="danger" onClick={this.delete}>탈퇴하기</Button>
                         <Button color="secondary" onClick={this.toggle}>취소</Button>
                     </ModalFooter>
                 </Modal>
-                {/* <Button color="danger" onClick={this.toggle}>탈퇴하기</Button> */}
             </div>
         );
     }
